Add theme provider to decorator
