+++ packages/ui-components/src/HorizontalLoader/index.stories.tsx
@@ -1,8 +1,7 @@
 import React from "react";
-import styled, { ThemeProvider } from "styled-components";
+import styled from "styled-components";
 
 import { HorizontalLoader } from ".";
-import { themes } from "../themes";
 
 export default { title: "Horizontal Loader" };
 
@@ -11,10 +10,8 @@ const Wrapper = styled.div`
 `;
 
 export const Default = () => (
-  <ThemeProvider theme={themes.default}>
-    <Wrapper>
-      <HorizontalLoader />
-      <div>Loading...</div>
-    </Wrapper>
-  </ThemeProvider>
+  <Wrapper>
+    <HorizontalLoader />
+    <div>Loading...</div>
+  </Wrapper>
 );
